removed all pyhook usage, replaced by punput tp resolve exit code error
