+++ tests/integration/collecting_processing/KeyboardTest.py
@@ -40,8 +40,9 @@ class KeyboardTest(unittest.TestCase):
         processor.join()
         features = self.keyboard_processor.features
         logging.debug(time.time() - st)
+
         self.assertAlmostEqual(features['typing_speed'], 3.8, delta=0.5)
-        self.assertAlmostEqual(features['active_typing_speed'], 5.4285, delta=0.5)
+        self.assertAlmostEqual(features['active_typing_speed'], 7.3, delta=0.5)
         self.assertAlmostEqual(features['average_press_duration'], 0.04, delta=0.5)
         self.assertAlmostEqual(features['average_down_to_down'], 0.08, delta=0.5)
         self.assertEqual(features['regular_press_count'], 16)
@@ -49,7 +50,7 @@ class KeyboardTest(unittest.TestCase):
         self.assertEqual(features['space_counter'], 3)
         self.assertEqual(features['error_corrections'], 0)
         self.assertEqual(features['mode_key'], ord('e'.upper()))
-        self.assertAlmostEqual(features['idle_time'], 1.5, delta=0.5)
+        self.assertAlmostEqual(features['idle_time'], 2.5, delta=0.5)
         self.assertEqual(features['unique_events'], 10)
 
         # database
