+++ tests/integration/collecting_processing_db/KeyboardTest.py
@@ -78,4 +78,4 @@ class KeyboardTest(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-    # hello from the test
+    #hello from the test
