+++ tests/processing/KeyboardProcessorTest.py
@@ -27,7 +27,7 @@ class KeyboardProcessorTest(unittest.TestCase):
         logging.debug(time.time() - st)
 
         self.assertAlmostEqual(features['typing_speed'], 3.8, delta=0.5)
-        self.assertAlmostEqual(features['active_typing_speed'], 5.4285, delta=0.5)
+        self.assertAlmostEqual(features['active_typing_speed'], 7.38, delta=0.5)
         self.assertAlmostEqual(features['average_press_duration'], 0.04, delta=0.5)
         self.assertAlmostEqual(features['average_down_to_down'], 0.08, delta=0.5)
         self.assertEqual(features['regular_press_count'], 16)
@@ -35,7 +35,7 @@ class KeyboardProcessorTest(unittest.TestCase):
         self.assertEqual(features['space_counter'], 3)
         self.assertEqual(features['error_corrections'], 0)
         self.assertEqual(features['mode_key'], ord('e'.upper()))
-        self.assertAlmostEqual(features['idle_time'], 1.5, delta=0.5)
+        self.assertAlmostEqual(features['idle_time'], 2.5, delta=0.5)
         self.assertEqual(features['unique_events'], 10)
 
         data_handler = KeyboardDataHandlerStub(name="processed")
