+++ tests/test_data/mouse/processed_data/processed_mouse.json
@@ -1,34 +1,1 @@
-{
-  "right_click_count": 1,
-  "left_click_count": 4,
-  "scroll_speed": 0,
-  "double_click_count": 2,
-  "cursor_x_distance": 0,
-  "cursor_y_distance": 0,
-  "average_momentary_speed_x": 0,
-  "average_momentary_speed_y": 0,
-  "average_speed_x": 0.0,
-  "average_speed_y": 0.0,
-  "average_active_speed_x": 0.0,
-  "average_active_speed_y": 0.0,
-  "average_cursor_angle": 0,
-  "std_cursor_angle": 0,
-  "Turn_0_45": 0,
-  "Turn_45_90": 0,
-  "Turn_90_135": 0,
-  "Turn_135_180": 0,
-  "Direction_S": 0,
-  "Direction_SW": 0,
-  "Direction_W": 0,
-  "Direction_NW": 0,
-  "Direction_N": 0,
-  "Direction_NE": 0,
-  "Direction_E": 0,
-  "Direction_SE": 0,
-  "Dist1": 0,
-  "Dist2": 0,
-  "Dist3": 0,
-  "idle_time": 3.9103403091430664,
-  "right_click_duration": 0.05222487449645996,
-  "left_click_duration": 0.06682801246643066
-}
+{"right_click_count": 1, "left_click_count": 4, "scroll_speed": 1517.859806225325, "double_click_count": 2, "cursor_x_distance": 0, "cursor_y_distance": 0, "average_momentary_speed_x": 0.0, "average_momentary_speed_y": 0.0, "average_speed_x": 0.0, "average_speed_y": 0.0, "average_active_speed_x": 0.0, "average_active_speed_y": 0.0, "average_cursor_angle": 0, "std_cursor_angle": 0, "Turn_0_45": 0, "Turn_45_90": 0, "Turn_90_135": 0, "Turn_135_180": 0, "Direction_S": 0, "Direction_SW": 0, "Direction_W": 0, "Direction_NW": 0, "Direction_N": 0, "Direction_NE": 0, "Direction_E": 0, "Direction_SE": 0, "Dist1": 0, "Dist2": 0, "Dist3": 0, "idle_time": 3.9078369140625, "right_click_duration": 0.06526970863342285, "left_click_duration": 0.06792831420898438}
