--- tests/test_data/pynput_kb/processed_data/pynput_processed_data.json
@@ -1,13 +1,0 @@
-{
-  "typing_speed": 3.8,
-  "active_typing_speed": 7.388737585964363,
-  "average_press_duration": 0.04660734732945761,
-  "average_down_to_down": 0.13293626433924624,
-  "regular_press_count": 16,
-  "punctuations_press_count": 0,
-  "space_counter": 3,
-  "error_corrections": 0,
-  "mode_key": 69,
-  "idle_time": 2.428518772125244,
-  "unique_events": 10
-}
