--- tests/test_data/pynput_kb/raw_data/pynput_raw_data.txt
@@ -1,39 +1,0 @@
-1613499472.076034
-0. {'Timestamp': 1613499473.5938091, 'Message': 256, 'KeyID': 72, 'Key': 'h'}
-1. {'Timestamp': 1613499473.6478965, 'Message': 257, 'KeyID': 72, 'Key': 'h'}
-2. {'Timestamp': 1613499473.7337437, 'Message': 256, 'KeyID': 69, 'Key': 'e'}
-3. {'Timestamp': 1613499473.7797747, 'Message': 257, 'KeyID': 69, 'Key': 'e'}
-4. {'Timestamp': 1613499473.8737335, 'Message': 256, 'KeyID': 76, 'Key': 'l'}
-5. {'Timestamp': 1613499473.9208908, 'Message': 257, 'KeyID': 76, 'Key': 'l'}
-6. {'Timestamp': 1613499474.0133643, 'Message': 256, 'KeyID': 76, 'Key': 'l'}
-7. {'Timestamp': 1613499474.0610554, 'Message': 257, 'KeyID': 76, 'Key': 'l'}
-8. {'Timestamp': 1613499474.1538477, 'Message': 256, 'KeyID': 79, 'Key': 'o'}
-9. {'Timestamp': 1613499474.2072086, 'Message': 257, 'KeyID': 79, 'Key': 'o'}
-10. {'Timestamp': 1613499474.294248, 'Message': 256, 'KeyID': 32, 'Key': 'space'}
-11. {'Timestamp': 1613499474.3415139, 'Message': 257, 'KeyID': 32, 'Key': 'space'}
-12. {'Timestamp': 1613499474.4340196, 'Message': 256, 'KeyID': 70, 'Key': 'f'}
-13. {'Timestamp': 1613499474.4812589, 'Message': 257, 'KeyID': 70, 'Key': 'f'}
-14. {'Timestamp': 1613499474.5741045, 'Message': 256, 'KeyID': 82, 'Key': 'r'}
-15. {'Timestamp': 1613499474.620349, 'Message': 257, 'KeyID': 82, 'Key': 'r'}
-16. {'Timestamp': 1613499474.7136626, 'Message': 256, 'KeyID': 79, 'Key': 'o'}
-17. {'Timestamp': 1613499474.7610757, 'Message': 257, 'KeyID': 79, 'Key': 'o'}
-18. {'Timestamp': 1613499474.852188, 'Message': 256, 'KeyID': 77, 'Key': 'm'}
-19. {'Timestamp': 1613499474.8986585, 'Message': 257, 'KeyID': 77, 'Key': 'm'}
-20. {'Timestamp': 1613499474.9913034, 'Message': 256, 'KeyID': 32, 'Key': 'space'}
-21. {'Timestamp': 1613499475.0399966, 'Message': 257, 'KeyID': 32, 'Key': 'space'}
-22. {'Timestamp': 1613499475.1330733, 'Message': 256, 'KeyID': 84, 'Key': 't'}
-23. {'Timestamp': 1613499475.178741, 'Message': 257, 'KeyID': 84, 'Key': 't'}
-24. {'Timestamp': 1613499475.2707043, 'Message': 256, 'KeyID': 72, 'Key': 'h'}
-25. {'Timestamp': 1613499475.318576, 'Message': 257, 'KeyID': 72, 'Key': 'h'}
-26. {'Timestamp': 1613499475.412328, 'Message': 256, 'KeyID': 69, 'Key': 'e'}
-27. {'Timestamp': 1613499475.4587145, 'Message': 257, 'KeyID': 69, 'Key': 'e'}
-28. {'Timestamp': 1613499475.55446, 'Message': 256, 'KeyID': 32, 'Key': 'space'}
-29. {'Timestamp': 1613499475.6033313, 'Message': 257, 'KeyID': 32, 'Key': 'space'}
-30. {'Timestamp': 1613499475.695083, 'Message': 256, 'KeyID': 84, 'Key': 't'}
-31. {'Timestamp': 1613499475.7429569, 'Message': 257, 'KeyID': 84, 'Key': 't'}
-32. {'Timestamp': 1613499475.8372667, 'Message': 256, 'KeyID': 69, 'Key': 'e'}
-33. {'Timestamp': 1613499475.8901243, 'Message': 257, 'KeyID': 69, 'Key': 'e'}
-34. {'Timestamp': 1613499475.9765766, 'Message': 256, 'KeyID': 83, 'Key': 's'}
-35. {'Timestamp': 1613499476.0254447, 'Message': 257, 'KeyID': 83, 'Key': 's'}
-36. {'Timestamp': 1613499476.1181405, 'Message': 256, 'KeyID': 84, 'Key': 't'}
-37. {'Timestamp': 1613499476.1660132, 'Message': 257, 'KeyID': 84, 'Key': 't'}
